v1.0/requirements.txt
pandas
sqlalchemy
mysql-connector-python
python-dotenv
openpyxl 